requirements.txt
flask==3.0.3
pytest==8.2.2
python-dotenv==1.0.1
